+++ src/constants.ts
@@ -38,6 +38,12 @@ export const DEFAULT_COMMENT = `
   [SVGO]: https://github.com/svg/svgo
 `;
 
+// Action outputs
+export const OUTPUT_NAME_DID_OPTIMIZE = "DID_OPTIMIZE";
+export const OUTPUT_NAME_OPTIMIZED_COUNT = "OPTIMIZED_COUNT";
+export const OUTPUT_NAME_SKIPPED_COUNT = "SKIPPED_COUNT";
+export const OUTPUT_NAME_SVG_COUNT = "SVG_COUNT";
+
 // File encodings
 export const BASE64 = "base64";
 export const UTF8 = "utf-8";