+++ test/schedule.test.ts
@@ -19,6 +19,10 @@ import {
   INPUT_NAME_REPO_TOKEN,
   GIT_OBJECT_TYPE_DIR,
   GIT_OBJECT_TYPE_FILE,
+  OUTPUT_NAME_DID_OPTIMIZE,
+  OUTPUT_NAME_OPTIMIZED_COUNT,
+  OUTPUT_NAME_SKIPPED_COUNT,
+  OUTPUT_NAME_SVG_COUNT,
   STATUS_ADDED,
 } from "../src/constants";
 import main from "../src/events/schedule";
@@ -101,6 +105,7 @@ beforeEach(() => {
   core.error.mockClear();
   core.info.mockClear();
   core.setFailed.mockClear();
+  core.setOutput.mockClear();
   core.warning.mockClear();
 
   encoder.decode.mockClear();
@@ -254,6 +259,80 @@ describe("Configuration", () => {
     expect(encoder.encode).toHaveBeenCalledTimes(1);
     expect(githubAPI.createBlob).toHaveBeenCalledTimes(1);
     expect(githubAPI.commitFiles).toHaveBeenCalledTimes(1);
+  });
+
+});
+
+describe("Outputs", () => {
+
+  const OUTPUTS_COUNT = 4;
+
+  test(`${OUTPUT_NAME_DID_OPTIMIZE} is set to "true"`, async () => {
+    const getContentsMock = mockGetContentsForFiles([fooFilePath]);
+    client.repos.getContent.mockImplementation(getContentsMock);
+
+    await main(client, config, svgo);
+
+    expect(core.setOutput).toHaveBeenCalledTimes(OUTPUTS_COUNT);
+    expect(core.setOutput).toHaveBeenCalledWith(
+      OUTPUT_NAME_DID_OPTIMIZE,
+      "true",
+    );
+  });
+
+  test(`${OUTPUT_NAME_DID_OPTIMIZE} is set to "false"`, async () => {
+    const getContentsMock = mockGetContentsForFiles([]);
+    client.repos.getContent.mockImplementation(getContentsMock);
+
+    await main(client, config, svgo);
+
+    expect(core.setOutput).toHaveBeenCalledTimes(OUTPUTS_COUNT);
+    expect(core.setOutput).toHaveBeenCalledWith(
+      OUTPUT_NAME_DID_OPTIMIZE,
+      "false",
+    );
+  });
+
+  test(`${OUTPUT_NAME_OPTIMIZED_COUNT} is set correctly`, async () => {
+    const getContentsMock = mockGetContentsForFiles([fooFilePath, barFilePath]);
+    client.repos.getContent.mockImplementation(getContentsMock);
+
+    await main(client, config, svgo);
+
+    expect(core.setOutput).toHaveBeenCalledTimes(OUTPUTS_COUNT);
+    expect(core.setOutput).toHaveBeenCalledWith(
+      OUTPUT_NAME_OPTIMIZED_COUNT,
+      "2",
+    );
+  });
+
+  test.each([
+    [[fooFilePath, barFilePath], 0],
+    [[optimizedFilePath], 1],
+  ])(`${OUTPUT_NAME_SKIPPED_COUNT} is set correctly`, async (files, count) => {
+    const getContentsMock = mockGetContentsForFiles(files);
+    client.repos.getContent.mockImplementation(getContentsMock);
+
+    await main(client, config, svgo);
+
+    expect(core.setOutput).toHaveBeenCalledTimes(OUTPUTS_COUNT);
+    expect(core.setOutput).toHaveBeenCalledWith(
+      OUTPUT_NAME_SKIPPED_COUNT,
+      `${count}`,
+    );
+  });
+
+  test(`${OUTPUT_NAME_SVG_COUNT} is set correctly`, async () => {
+    const getContentsMock = mockGetContentsForFiles([fooFilePath, barFilePath]);
+    client.repos.getContent.mockImplementation(getContentsMock);
+
+    await main(client, config, svgo);
+
+    expect(core.setOutput).toHaveBeenCalledTimes(OUTPUTS_COUNT);
+    expect(core.setOutput).toHaveBeenCalledWith(
+      OUTPUT_NAME_SVG_COUNT,
+      "2",
+    );
   });
 
 });